[release-branch.go1.22] runtime: x_cgo_getstackbound: initialize pthread attr
In glibc versions older than 2.32 (before commit 4721f95058),
pthread_getattr_np does not always initialize the `attr` argument,
and when it fails, it results in a NULL pointer dereference in
pthread_attr_destroy down the road.

This is the simplest way to avoid this, and an alternative to CL 585019.

Updates #65625.

(cherry picked from commit 771ff5aabad99ae2151c05c7e8a73f1153365586)

Change-Id: I2bcd4cfcd9a79813147c1e8378d1534c32924952