cmd/go: set the GoVersion for files listed on the commandline
Fixes: #65612
Fixes: #66092

The decision we came to is to default to the containing module's go
version. The containing module means the module that contains the
specified file. If there is no containing module, we will default to using the
local go version.

The compiler will need to adjust for this behavior change as it always
used the local go version for files listed on the commandline.

This CL is expected to break some existing compiler tests in order to
allow for the change in behavior.

Tested: Ran go test cmd/go
Cq-Include-Trybots: luci.golang.try:gotip-linux-amd64-longtest,gotip-windows-amd64-longtest
Change-Id: Ia8b0a0b1905aef394a0c76cd0015e936c7e89a6e