cmd/go: set the GoVersion for files listed on the commandline with vet
Fixes: #65612
Fixes: #66092

The decision we came to is to default to the containing module's go
version. The containing module means the module that contains the
specified file. If there is no containing module, we will default to using the
local go version.

The compiler will need to adjust for this behavior change as it always
used the local go version for files listed on the commandline.

This CL is expected to break some existing compiler tests in order to
allow for the change in behavior.

With this change, we may want to revisit how we create a module for
command line arguments in src/cmd/go/internal/modload/init.go
Currently we always set the go version to the local go version. I added
a TODO.

Tested: Ran go test cmd/go
Cq-Include-Trybots: luci.golang.try:gotip-linux-amd64-longtest,gotip-windows-amd64-longtest
Change-Id: Iad106db736e9231091dae48d315311d4d7da9dbf