cmd/compile: remove quoting in favor of clearer prose in error messages
To address issue #65790 (confusing error messages), quoting of names
was introduced in error messages. It solved the issue at hand at the
cost of extra text (the quotes) in error messages; it also introduced
an inconsistent use of quotes.

This CL removes the quoting again in favor or adding a qualifying noun
(such as "name", "label", "package", etc.) before the offending name
where needed.

For instance, instead of

        invalid argument to `max'

we now say

        invalid argument to built-in max

This change touches a lot of files but only affects error messages.

Fixes #67685.

Change-Id: I95435b388f92cade316e2844d59ecf6953b178bc