os/exec: revert Windows-related LookPath changes
This reverts the three most recent CLs about Windows PATH lookups:

 - CL 528038 "avoid calling LookPath in cmd.Start for resolved paths"
 - CL 528037 "fix edge cases in Windows PATH resolution"
 - CL 527820 "avoid writing to Cmd.Path in Cmd.Start on Windows"

The next CL in the stack adds TestAbsPathExec to test for the Go 1.21 behavior.
With the three CLs reverted, the test passes.
(With fewer CLs reverted, the test does not pass.)

A future CL can attempt to replay some of the changes for Go 1.23,
but for Go 1.22 we can just backport these reverts and we will be
no worse off than in Go 1.21.

For #66586.

Change-Id: I059034f09ec30c2009fc28cbe8bc96b22962ed98