cmd/go/internal/modload: respect overlays when loading go.work files
Before this change, we didn't initialize the overlays in the fsys
package or use the fsys logic to read the files, so overlays were not
respected for go.work files. Initialize fsys before loading the go.work
file (initialization is idempotent) and use fsys.OverlayPath to get the
true path of the go.work file before reading it.

Change-Id: Ic40bcbb483a16c5d4dd1d896306ea99a16f370f3