time: remove the leading 0 from the day in time.RFC1123{,Z}
When using time.RFC1123Z to parse the date header value out of an email,
an error is returned for dates that occur in the first 9 days of a month.
This is because the format strings for RFC 1123 defined in the time
package indicate that the day should be prefixed with a leading 0.

Reading the spec, the line that talks about it seems to indicate that
2-digit days are only allowed if the leading digit is a '1':

  `date = 1*2DIGIT month 2*4DIGIT`

Fixes #67887

Change-Id: Ie7ee40d94da2c8c0417957e8b89f9987314949c8
GitHub-Last-Rev: 8621813b86c8171c17700f8f27cc50644f7a5f43
GitHub-Pull-Request: golang/go#67888