os: rewrite TestChtimesWithZeroTimes
There are a number of problems with TestChtimesWithZeroTimes case:

 - test cases are interdependent (former cases influence the latter ones),
   making the test using too many different times and a bit hard to read;
 - time is changed forward not backword which is somewhat more racy;
 - the "omit none" test case is redundant (same as in TestChtimes);
 - if the test has failed, it hard to see which case is failing;
 - atime exclusion caused by mount noatime flag applies to all unix
   (but is only needed for netbsd, same as in testChtimes);
 - mountinfo is being read from /bin/mounts (obviously wrong path);
 - the atime check is done twice for default ("unix") case.

Let's rewrite this, fixing all these issues.

Also, since CL 589496 this should work on DragonFly BSD. For the most
part, that is -- when both times are omitted, it doesn't work due to
a kernel bug, so add an exception about it.

Change-Id: If9020256ca920b4db836a1f0b2e055b5fce4a552