cmd/go: call telemetry.MaybeChild at start of go command
Call the new telemetry.MaybeChild function at the start of the go
command so that the child process logic can be run immediately without
running toolchain selection if this is the child process. The
StartWithUpload function in the telemetry shim package has been renamed
to MaybeParent because that's its actual effective behavior in cmd/go,
the only place it's called: it won't run as the child because MaybeChild
has already been called and would have run as the child if the program
was the telemetry child, and it won't open counters because
telemetry.Start has been called. It might make sense to add an actual
telemetry.MaybeParent function that doesn't try to start the child or
open counters to make things a little simpler.

Change-Id: Ie81e2418af85cef18ec41f75db66365f6597b8b1