cmd/internal/obj: Add a new handler for register printing from Addr
The purpose of this change is to provide access to the whole Addr
structure when printing a register. Currently each architecture is
allocated a range of a few thousand numbers to represent registers.
Adding a callback that can be registered to interpret any arbitrary
obj.Addr as a register would allow any backend to override the generic
register printing logic with their own, so they can interpret the Addr
freely. This would allow the backend to use more bits to represent a
register that wouldn't normally be supported by their RConv callback.
The callback is accessed from a lookup table, keyed by the architecture
family stored in the assembler context.

The second part of this change registers a handler for the ARM64 backend
and moves an ARM64 specific edge case into the handler.

Updates #68509

Change-Id: Iab15d7ada2057d43cb3afc556a1e95ef01f53b32