crypto/subtle: add DIT closure
Add a new function, WithDataIndependentTiming,
which takes a function as an argument, and
encloses it with calls to set/unset the DIT PSTATE
bit on Arm64.

Since DIT is OS thread-local, for the duration of
the execution of WithDataIndependentTiming, we
lock the goroutine to the OS thread, using
LockOSThread. For long running operations, this is
likely to not be performant, but we expect this to
be tightly scoped around cryptographic operations
that have bounded execution times.

If locking to the OS thread turns out to be too
slow, another option is to add a bit to the g
state indicating if a goroutine has DIT enabled,
and then have the scheduler enable/disable DIT
when scheduling a g.

Fixes #66450
Updates #49702

Change-Id: I78eda691ff9254b0415f2b54770e5850a0179749