cmd/internal/cov: close counter data files eagerly
When reading the counter dat files in a given pod, close the
underlying file after each one is read, as opposed to using a deferred
close in the loop (which will close them all at the end of the
function). Avoids running into "too many open files" when processing
large clumps of counter data files.

Fixes #68468.

Change-Id: Ic1fe1d36c44d3f5d7318578cd18d0e65465d71d9