text/template: fix doc spacing
This doc syntax:

```md
- The name of a field of the data, which must be a struct, preceded
  by a period, such as
		.Field
```

was producing this formatting in the terminal and pkg.go.dev:

```
The name of a field of the data, which must be a struct, preceded by a period, such as .Field The result is the value of the field.
```

This change makes the formatting more aligned with the intent, and doesn't have bad punctuation:

```
The name of a field of the data, which must be a struct, preceded by a period, such as

.Field

The result is the value of the field.
```

There doesn't appear to be a way to have pre-formatted content inside a list in Go doc, sadly.

Change-Id: I4550ce3135b3cd675010a3701618e9350f282453
GitHub-Last-Rev: d2e52f14968621e9913dd66ef6b26348012c83b3
GitHub-Pull-Request: golang/go#68519