all: support type parameterized aliases in export data
This adds a new ObjGenericAlias for type parameterized aliases
in the unifed IR export data, ObjGenericAlias. This extends
ObjAlias by additionally writing the type parameters of the
Alias.

This is currently only emitted when the AliasTypeParams
experiment is enabled.

We cannot simply extend ObjAlias with type parameters as
this would break both old tools that do not anticipate type
parameters in the encoding, and updated tools that do
anticipate type parameters in the encoding that are not
emitted by older compilers. For example, an x/tools/go/package
based tools compiled with a new version of Go with type parameters
still needs to be able to read the unified IR created by the Go 1.22
compiler for a package.

Updates #68526

Change-Id: Ie43641c6172d9639b86530eaea76dee94c5c976d