WIP: cmd/relnote: add test to check links for problems
DO NOT REVIEW: This started out as a checkpoint that explored reusing
as much of x/website/cmd/golangorg's TestAll as possible to check the
links in the doc/next content.

It's nice that it's viable to reuse x/build/relnote which is already
vendored into cmd module to generate the release notes draft HTML,
but unfortunately this is the first use of x/net/html in this repo.

An alternative I considered was doing a simpler string-based check
for URL problems, but that might miss problems that x/website test
catches so it's not ideal.

DO NOT SUBMIT: It should work better to reuse x/website tests by
triggering them as a trybot when *.md files are modified, instead of
copying the test here.

For #68575.

Change-Id: Ic71054580401d5eac2fe1c2173e7dad55f189bda