WIP: cmd/relnote: add test to check links for problems
DO NOT REVIEW: This is an incomplete checkpoint that explored reusing
as much of x/website/cmd/golangorg's TestAll as possible to check the
links in the doc/next content.

It's nice that it's viable to reuse x/build/relnote which is already
vendored into cmd module to generate the release notes draft HTML,
but unfortunately this is the first use of x/net/html in this repo.

An alternative I'll consider/explore is doing a simpler string-based
check for URL problems. This is mailed as a draft in the meantime.

DO NOT SUBMIT: The CL is not finished.

For #68575.

Change-Id: Ic71054580401d5eac2fe1c2173e7dad55f189bda