[release-branch.go1.23] cmd/compile: avoid compiler panic when instantiating imported generic type alias
When GOEXPERIMENT=aliastypeparams is set, type aliases may have type
parameters. The compiler export data doesn't export the type parameter
information yet, which leads to a panic when a client package tries to
instantiate such an alias type.

This CL prevents the panic and instead produces an invalid type for
type parameters of such generic alias types. As a result, a user will
see an error message about the alias type not being generic. This is
not correct, of course, but it is better than a panic. The change is
only in effect if the GOEXPERIMENT is enabled.

For #68526.

Change-Id: I24737b153a7e2f9b705cd29a5b70b2b9e808dffc